refractor: break into smaller scripts
